添加选中角标功能
将选中角标的方法暴露出来，可以设置选中第几个。

+++ QZPageSwitchDemo/QZPageSwitchDemo/QZPageSwitch.m
@@ -99,7 +99,7 @@
         [self setSelectedIndex:index animated:YES];
     }
 }
-- (void)setSelectedIndex:(int)selectedIndex animated:(BOOL)animated {
+- (void)setSelectedIndex:(NSInteger)selectedIndex animated:(BOOL)animated {
     if (selectedIndex <= self.titleLabels.count) {
         BOOL catchHalfSwitch = NO;
         if (self.selectedIndex == selectedIndex) {
